Separate out api from twitch client
